louder CTA buttons now for single project modal
