+++ src/components/SingleProjectModal.jsx
@@ -2,10 +2,11 @@
 /* eslint-disable react/prop-types */
 /* eslint-disable max-len */
 import React, {
-  useState, useRef, useEffect,
+  useState, useRef, useEffect, useContext,
 } from 'react';
 import axios from 'axios';
 import {
+  ButtonGroup,
   Chip,
   Grid,
   Button,
@@ -19,22 +20,81 @@ import {
   IconButton,
   Container,
   Box,
+  ListItemIcon,
+  ListItemText,
 } from '@mui/material';
 import BACKEND_URL from '../supportFunctions.js';
 import Iconify from './Iconify.jsx';
+import { UserContext } from './UserContext.jsx';
 
 import CircularIndeterminate from './CircularIndeterminate.jsx';
 import UserMoreMenu from './UserMoreMenu.jsx';
 import HorizontalStepper from './HorizontalStepper.jsx';
+import SingleProjectKanbanModal from './SingleProjectKanbanModal.jsx';
 
 export default function SingleProjectModal({ rowContent, setJustSubmitted }) {
+  const { user } = useContext(UserContext);
   const [open, setOpen] = useState(false);
   const [scroll] = useState('paper');
 
   const [showLoading, setShowLoading] = useState(true);
+  const [showJoinButton, setShowJoinButton] = useState(false);
 
   const [usersList, setUsersList] = useState([]);
   const [skillsList, setSkillsList] = useState([]);
+
+  const checkIfUserIsAlreadyEnrolledHere = () => {
+    if (usersList.some((name) => name === user.name)) {
+      return true;
+    } return false;
+  };
+
+  async function addUserToProject() {
+    try {
+      const enrolUserToProjectObject = {
+        userId: user.id,
+        projectId: rowContent.id,
+      };
+      const enrolUser = await axios.post(`${BACKEND_URL}/project/enrol-user/${user.id}`, enrolUserToProjectObject);
+      console.log(enrolUser);
+    } catch (error) {
+      console.log(error);
+    }
+  }
+  const handleAddUserToProject = () => {
+    console.log('USER ID:', user.id, 'ADDED PROJECT', rowContent.id);
+    setOpen(false);
+    addUserToProject();
+    setJustSubmitted(true);
+  };
+
+  const checkIfProjectFull = () => {
+    const engineersEnrolled = Number(rowContent.user_projects.length);
+    const engineersRequired = Number(rowContent.noEngineersRequired);
+    if ((engineersEnrolled / engineersRequired) === 1) return true;
+    return false;
+  };
+
+  const checkDateValid = () => {
+    const date = (rowContent.deliveryDeadline.slice(0, 10)).value;
+    const varDate = new Date(date); // dd-mm-YYYY
+    const today = new Date();
+    today.setHours(0, 0, 0, 0);
+    // check if the project deadline date is bigger than today, if yes, TRUE
+    if (varDate >= today) {
+      console.log('checkDateValid!');
+      return true;
+    }
+    return false;
+  };
+
+  useEffect(() => {
+    if (rowContent.stage === 'sourcing' && !checkDateValid() && !checkIfProjectFull() && user.accountType === 'engineer' && !checkIfUserIsAlreadyEnrolledHere()) {
+      setShowJoinButton(true);
+    } else {
+      console.log('this user cannot join this project.');
+    }
+  }, []);
 
   async function getUsersAndSkillsForThisProject() {
     try {
@@ -98,7 +158,7 @@ export default function SingleProjectModal({ rowContent, setJustSubmitted }) {
               </Typography>
             </Grid>
             <Grid item xs={1}>
-              {!showLoading && <UserMoreMenu rowContent={rowContent} usersList={usersList} setJustSubmitted={setJustSubmitted} />}
+              {!showLoading && user.accountType === 'manager' && <UserMoreMenu rowContent={rowContent} setJustSubmitted={setJustSubmitted} />}
             </Grid>
           </Grid>
 
@@ -272,12 +332,39 @@ export default function SingleProjectModal({ rowContent, setJustSubmitted }) {
               </Container>
             </DialogContentText>
           </DialogContent>
-
         )}
         <DialogActions alignItems="center">
 
-          <Button onClick={handleClose}>Close</Button>
-
+          {!showLoading && (
+          <ButtonGroup variant="contained" aria-label="outlined button group">
+            {showJoinButton ? (
+              <Button color="success" onClick={(e) => handleAddUserToProject(e)}>
+                <ListItemIcon color="success">
+                  <Iconify
+                    icon="fluent:arrow-join-20-regular"
+                    width={24}
+                    height={24}
+                  />
+                </ListItemIcon>
+                <ListItemText primary="Join Project" primaryTypographyProps={{ variant: 'body2' }} />
+              </Button>
+            ) : (
+              <Button sx={{ color: 'text.secondary' }} disabled>
+                <ListItemIcon>
+                  <Iconify
+                    icon="fluent:arrow-join-20-regular"
+                    width={24}
+                    height={24}
+                  />
+                </ListItemIcon>
+                <ListItemText primary="Join Project" primaryTypographyProps={{ variant: 'body2' }} />
+              </Button>
+            )}
+              {rowContent.stage === 'in-progress' && <SingleProjectKanbanModal projectId={rowContent.id} name={rowContent.name} data={rowContent.kanbanData} setJustSubmitted={setJustSubmitted} />}
+            {' '}
+            <Button onClick={handleClose}>Close</Button>
+          </ButtonGroup>
+          )}
         </DialogActions>
       </Dialog>
     </>
